requirements.txt
asgiref==3.8.1
colorama==0.4.6
coverage==7.8.0
Django==5.2.1
flake8==7.2.0
iniconfig==2.1.0
mccabe==0.7.0
packaging==25.0
pillow==11.2.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pyflakes==3.3.2
pytesseract==0.3.13
pytest==8.3.5
pytest-django==4.11.1
sqlparse==0.5.3
tzdata==2025.2
